changed the floatData structure to accommodate the extra param
The extra param added to the data was a column at the start of the page listing the name of the instrument. I increased the number of data elements to 27 and added an extra parameter "name" in the FloatData class that stores the name of the string. The application now builds successfully but it still has a few issues. I suspect that they have something to do with the Google Maps API key